Added listeners for image processing, sped up optical flow by reducing resolution

+++ app/src/main/java/com/robot/balancingandroid/Estimator.java
@@ -191,7 +191,7 @@ public class Estimator
         L = Mat.zeros(3, 3, CvType.CV_32FC1);
     }
 
-    public void onFlowChanged(float flow) {
+    public void onFlowChanged(double flow) {
         flowMes = flow;
         newFlowMes = true;
     }
